+++ src/MatrixMate.php
@@ -11,19 +11,23 @@
 namespace vaersaagod\matrixmate;
 
 use Craft;
+use craft\base\Element;
 use craft\base\Plugin;
+use craft\elements\Asset;
+use craft\elements\Category;
 use craft\elements\Entry;
+use craft\elements\GlobalSet;
+use craft\elements\User;
 use craft\helpers\Json;
-use craft\services\Categories;
 use craft\services\Fields;
-use craft\services\Globals;
 use craft\services\Plugins;
-use craft\services\Sections;
+use craft\web\Controller;
 
 use vaersaagod\matrixmate\assetbundles\matrixmate\MatrixMateAsset;
 use vaersaagod\matrixmate\services\MatrixMateService;
 use vaersaagod\matrixmate\models\Settings;
 
+use yii\base\ActionEvent;
 use yii\base\Event;
 
 /**
@@ -37,39 +41,48 @@ use yii\base\Event;
  */
 class MatrixMate extends Plugin
 {
-    // Static Properties
-    // =========================================================================
 
     /**
-     * @var MatrixMate
+     * @var Settings
+     * @see getSettings()
      */
-    public static $plugin;
-
-    // Public Properties
-    // =========================================================================
-
-    /**
-     * @var string
-     */
-    public $schemaVersion = '1.0.0';
-
-    // Public Methods
-    // =========================================================================
+    private Settings $_settings;
 
     /**
      * @inheritdoc
      */
     public function init()
     {
+
         parent::init();
-        self::$plugin = $this;
+
+        if (!$this->isInstalled) {
+            return;
+        }
 
         // Register services
         $this->setComponents([
             'matrixMate' => MatrixMateService::class,
         ]);
 
-        $this->addEventListeners();
+        // Clear the field config cache when something field related happens
+        foreach ([
+             Fields::EVENT_AFTER_SAVE_FIELD_LAYOUT => Fields::class,
+             Fields::EVENT_AFTER_DELETE_FIELD_LAYOUT => Fields::class,
+             Fields::EVENT_AFTER_SAVE_FIELD => Fields::class,
+             Fields::EVENT_AFTER_DELETE_FIELD => Fields::class,
+        ] as $event => $class) {
+            Event::on($class, $event, [$this->matrixMate, 'clearFieldConfigCache']);
+        }
+
+        // Defer further initialisation to after plugins have loaded, and only for CP web requests
+        if (Craft::$app->getRequest()->getIsCpRequest() && !Craft::$app->getRequest()->getIsConsoleRequest()) {
+            Event::on(
+                Plugins::class,
+                Plugins::EVENT_AFTER_LOAD_PLUGINS,
+                [$this, 'onAfterLoadPlugins']
+            );
+        }
 
         Craft::info(
             Craft::t(
@@ -82,170 +95,107 @@ class MatrixMate extends Plugin
     }
 
     /**
-     * @throws \yii\base\InvalidConfigException
+     * @return void
      */
-    public function onAfterLoadPlugins()
+    public function onAfterLoadPlugins(): void
     {
 
-        $request = Craft::$app->getRequest();
-        $user = Craft::$app->getUser();
-
-        if (!$this->isInstalled || !$request->getIsCpRequest() || $request->getAcceptsJson() || $request->getIsConsoleRequest() || !$user->checkPermission('accessCp')) {
+        if (!Craft::$app->getUser()->checkPermission('accessCp')) {
             return;
         }
 
-        // Get the field config
-        $fieldConfig = MatrixMate::$plugin->matrixMate->getFieldConfig();
-        if (!$fieldConfig) {
-            // Bail early if no field config
-            return;
-        }
+        // Entries, assets and categories
+        Craft::$app->getView()->hook('cp.elements.edit', function (array $context) {
+            $this->registerAssetBundleForElement($context['element'] ?? null);
+        });
 
-        $segments = $request->getSegments();
-        if (empty($segments)) {
-            return;
-        }
+        // Users
+        Craft::$app->getView()->hook('cp.users.edit', function (array $context) {
+            $this->registerAssetBundleForElement($context['user'] ?? null);
+        });
 
-        $context = '*';
-        $isEntryVersion = false;
+        // Global sets
+        Craft::$app->getView()->hook('cp.globals.edit', function (array $context) {
+            $this->registerAssetBundleForElement($context['globalSet'] ?? null);
+        });
 
-        if (\count($segments) >= 3 && $segments[0] === 'entries') {
-            $entryType = null;
-            if ($segments[2] === 'new') {
-                // New entry – check if there's a (valid) typeId param in the URL
-                $typeIdParam = (int)$request->getParam('typeId', null);
-                if (!$typeIdParam || !$entryType = Craft::$app->getSections()->getEntryTypeById($typeIdParam)) {
-                    // Nope, use the first one for the current section
-                    $section = Craft::$app->getSections()->getSectionByHandle($segments[1]);
-                    $entryType = $section ? $section->getEntryTypes()[0] : null;
+        // Slideouts
+        Event::on(
+            Controller::class,
+            \yii\base\Controller::EVENT_BEFORE_ACTION,
+            function (ActionEvent $event) {
+                $action = $event->action;
+                if ($action->id !== 'get-editor-html') {
+                    return;
                 }
-            } else {
-                // Existing entry – get the entry and use its entry type
-                $entryId = (int)\explode('-', $segments[2])[0];
-                if ($entryId) {
-                    $siteHandle = $request->getParam('site', null);
-                    $site = $siteHandle ? Craft::$app->getSites()->getSiteByHandle($siteHandle) : null;
-                    $siteId = $site ? $site->id : '*';
-                    $entryQuery = Entry::find()->siteId($siteId)->status(null);
-                    $draftId = (int)$request->getParam('draftId', null);
-                    if ($draftId) {
-                        // They're looking at a bona fide draft
-                        $draftsQuery = clone $entryQuery;
-                        $entry = $draftsQuery->draftId($draftId)->one();
-                    } else if (\version_compare(Craft::$app->getVersion(), '3.7.0', '>=')) {
-                        // If Craft 3.7+, look for provisional draft
-                        $provisionalDraftsQuery = clone $entryQuery;
-                        $entry = $provisionalDraftsQuery->provisionalDrafts()->draftCreator($user->identity)->draftOf($entryId)->one();
-                    }
-                    $entry = $entry ?? $entryQuery->id($entryId)->one();
-                    if ($entry) {
-                        $entryType = $entry->getType();
-                    }
+                $request = $action->controller->request;
+                $elementId = (int)$request->getBodyParam('elementId');
+                $elementType = $request->getBodyParam('elementType');
+                $siteId = (int)$request->getBodyParam('siteId');
+                if ($elementId && $elementType && $siteId) {
+                    $element = Craft::$app->getElements()->getElementById($elementId, $elementType, $siteId);
+                    $this->registerAssetBundleForElement($element);
                 }
             }
-            if ($entryType) {
-                $context = "entryType:{$entryType->id}";
-            }
-            if (($segments[3] ?? null) === 'versions' || !!$request->getParam('revisionId')) {
-                $isEntryVersion = true;
-            }
-        } else if (\count($segments) >= 3 && $segments[0] === 'categories') {
-            if ($group = Craft::$app->getCategories()->getGroupByHandle($segments[1])) {
-                $context = "categoryGroup:{$group->id}";
-            }
-        } else if (\count($segments) >= 2 && $segments[0] === 'globals') {
-            if ($globalSet = Craft::$app->getGlobals()->getSetByHandle($segments[\count($segments) - 1])) {
-                $context = "globalSet:{$globalSet->id}";
-            }
-        } else if ($segments[0] == 'myaccount' || (\count($segments) === 2 && $segments[0] === 'users')) {
-            $context = 'users';
-        }
-
-        $settings = [
-            'context' => $context,
-            'fieldsConfig' => $fieldConfig,
-            'isEntryVersion' => $isEntryVersion,
-            'isCraft34' => \version_compare(Craft::$app->getVersion(), '3.4.0', '>='),
-            'isCraft35' => \version_compare(Craft::$app->getVersion(), '3.5.0-RC1', '>='),
-        ];
-
-        $view = Craft::$app->getView();
-        $view->registerAssetBundle(MatrixMateAsset::class);
-
-        $view->registerJs('if (Craft && Craft.MatrixMate) { new Craft.MatrixMate(' . Json::encode($settings, JSON_UNESCAPED_UNICODE) . '); }');
+        );
 
     }
 
     /**
-     *  Clear the field config cache when something field related happens
+     * @return Settings
      */
-    public function onAfterSaveFieldContext()
+    public function getSettings(): Settings
     {
-        MatrixMate::$plugin->matrixMate->clearFieldConfigCache();
+        if (!isset($this->_settings)) {
+            $this->_settings = $this->createSettingsModel();
+        }
+        return $this->_settings;
     }
 
     // Protected Methods
     // =========================================================================
 
     /**
-     * @inheritdoc
+     * @return Settings
      */
-    protected function createSettingsModel()
+    protected function createSettingsModel(): Settings
     {
         return new Settings();
     }
 
     /**
-     *
+     * @param Element|null $element
+     * @return void
+     * @throws \yii\base\InvalidConfigException
      */
-    protected function addEventListeners()
+    protected function registerAssetBundleForElement(?Element $element): void
     {
-        Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_LOAD_PLUGINS,
-            [$this, 'onAfterLoadPlugins']
-        );
+        if (!$element || !$fieldsConfig = MatrixMate::getInstance()->matrixMate->getFieldConfig()) {
+            return;
+        }
 
-        Event::on(
-            Fields::class,
-            Fields::EVENT_AFTER_SAVE_FIELD_LAYOUT,
-            [$this, 'onAfterSaveFieldContext']
-        );
+        $context = '*';
 
-        Event::on(
-            Fields::class,
-            Fields::EVENT_AFTER_DELETE_FIELD,
-            [$this, 'onAfterSaveFieldContext']
-        );
+        if ($element instanceof Entry) {
+            $context = "entryType:$element->typeId";
+        } else if ($element instanceof Category) {
+            $context = "categoryGroup:$element->groupId";
+        } else if ($element instanceof Asset) {
+            $context = "volume:$element->volumeId";
+        } else if ($element instanceof GlobalSet) {
+            $context = "globalSet:$element->handle";
+        } else if ($element instanceof User) {
+            $context = 'users';
+        }
 
-        Event::on(
-            Fields::class,
-            Fields::EVENT_AFTER_SAVE_FIELD,
-            [$this, 'onAfterSaveFieldContext']
-        );
+        $settings = [
+            'context' => $context,
+            'fieldsConfig' => $fieldsConfig,
+            'isRevision' => (bool)$element->revisionId,
+        ];
 
-        Event::on(
-            Sections::class,
-            Sections::EVENT_AFTER_SAVE_ENTRY_TYPE,
-            [$this, 'onAfterSaveFieldContext']
-        );
-
-        Event::on(
-            Sections::class,
-            Sections::EVENT_AFTER_SAVE_SECTION,
-            [$this, 'onAfterSaveFieldContext']
-        );
-
-        Event::on(
-            Categories::class,
-            Categories::EVENT_AFTER_SAVE_GROUP,
-            [$this, 'onAfterSaveFieldContext']
-        );
-
-        Event::on(
-            Globals::class,
-            Globals::EVENT_AFTER_SAVE_GLOBAL_SET,
-            [$this, 'onAfterSaveFieldContext']
-        );
+        Craft::$app->getView()->registerAssetBundle(MatrixMateAsset::class, \yii\web\View::POS_END);
+        Craft::$app->getView()->registerJs('if (Craft && Craft.MatrixMate) { new Craft.MatrixMate(' . Json::encode($settings, JSON_UNESCAPED_UNICODE) . '); }', \yii\web\View::POS_END);
     }
+
 }